+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - calendar booking details to payment receipt
+- Display of images for Checkboxes and Radio button options
 
 ## [6.10.0] - 2024-11-27